address inconsistant data types between models and sql models

+++ src/api/models/sql_models/benthiclit.py
@@ -183,8 +183,8 @@ class BenthicLITObsSQLModel(BaseSUSQLModel):
     sample_unit_notes = models.TextField(blank=True)
 
     transect_number = models.PositiveSmallIntegerField()
-    transect_len_surveyed = models.PositiveSmallIntegerField(
-        verbose_name=_("transect length surveyed (m)")
+    transect_len_surveyed = models.DecimalField(
+        max_digits=4, decimal_places=1, verbose_name=_("transect length surveyed (m)")
     )
     reef_slope = models.CharField(max_length=50)
     length = models.PositiveSmallIntegerField()
@@ -336,8 +336,8 @@ class BenthicLITSUSQLModel(BaseSUSQLModel):
     sample_unit_notes = models.TextField(blank=True)
 
     transect_number = models.PositiveSmallIntegerField()
-    transect_len_surveyed = models.PositiveSmallIntegerField(
-        verbose_name=_("transect length surveyed (m)")
+    transect_len_surveyed = models.DecimalField(
+        max_digits=4, decimal_places=1, verbose_name=_("transect length surveyed (m)")
     )
     total_length = models.PositiveIntegerField()
     reef_slope = models.CharField(max_length=50)
